[spc][bbk] Add an origin/domain matcher for deleting BBKs
The matcher prevents arbitrary origins and domains from deleting all
BBKs from the Chrome profile.

Fixed: 519241232
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7899197
Auto-Submit: Darwin Yang <darwinyang@chromium.org>
Commit-Queue: Darwin Yang <darwinyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1643259}

+++ chrome/browser/browsing_data/chrome_browsing_data_remover_delegate.cc
@@ -695,11 +695,13 @@ void ChromeBrowsingDataRemoverDelegate::RemoveEmbedderData(
     }
 #endif  // !BUILDFLAG(IS_ANDROID)
 
-    if (payments::
-            BrowserBoundKeyDeleterService* browser_bound_key_deleter_service =
-                payments::BrowserBoundKeyDeleterServiceFactory::GetForProfile(
-                    profile_)) {
-      browser_bound_key_deleter_service->RemoveInvalidBBKs();
+    if (filter_builder->MatchesMostOriginsAndDomains()) {
+      if (payments::
+              BrowserBoundKeyDeleterService* browser_bound_key_deleter_service =
+                  payments::BrowserBoundKeyDeleterServiceFactory::GetForProfile(
+                      profile_)) {
+        browser_bound_key_deleter_service->RemoveInvalidBBKs();
+      }
     }
 
 #if BUILDFLAG(IS_CHROMEOS)
